Add bow endpoint and fix one issue regarding elements
Around 300 weapons' elements are still broken due to the issue described in GH-1.
Fixed an issue where a weapon's element was being stored as a single string/int. Seeing as weapons can have multiple elements, and that weapon defense is also included in this table, this lead to some important things being left out...